Fix bookmarks key error for direct access to xblock.
TNL-5003
