+++ common/lib/xmodule/xmodule/tests/test_vertical.py
@@ -1,9 +1,11 @@
 """
 Tests for vertical module.
 """
-
+import ddt
+from mock import Mock
 from fs.memoryfs import MemoryFS
 from xmodule.tests import get_test_system
+from xmodule.tests.helpers import StubUserService
 from xmodule.tests.xml import XModuleXmlImportTest
 from xmodule.tests.xml import factories as xml
 from xmodule.x_module import STUDENT_VIEW, AUTHOR_VIEW
@@ -41,6 +43,7 @@ class BaseVerticalBlockTest(XModuleXmlImportTest):
         self.default_context = {"bookmarked": False, "username": self.username}
 
 
+@ddt.ddt
 class VerticalBlockTestCase(BaseVerticalBlockTest):
     """
     Tests for the VerticalBlock.
@@ -54,11 +57,21 @@ class VerticalBlockTestCase(BaseVerticalBlockTest):
         self.assertIn('bookmarked', content)
         self.assertIn('show_bookmark_button', content)
 
-    def test_render_student_view(self):
+    @ddt.unpack
+    @ddt.data(
+        {'context': None},
+        {'context': {}}
+    )
+    def test_render_student_view(self, context):
         """
         Test the rendering of the student view.
         """
-        html = self.module_system.render(self.vertical, STUDENT_VIEW, self.default_context).content
+        self.module_system._services['bookmarks'] = Mock()  # pylint: disable=protected-access
+        self.module_system._services['user'] = StubUserService()  # pylint: disable=protected-access
+
+        html = self.module_system.render(
+            self.vertical, STUDENT_VIEW, self.default_context if context is None else context
+        ).content
         self.assertIn(self.test_html_1, html)
         self.assertIn(self.test_html_2, html)
         self.assert_bookmark_info_in(html)
